webkitgtk: fix reconfigure
| ln: failed to create symbolic link '/home/superandy/tmp/oe-core-glibc/work/cortexa7t2hf-neon-vfpv4-mortsgna-linux-gnueabi/webkitgtk/2.26.4-r0/recipe-sysroot-native/usr/bin/python': File exists

+++ meta/recipes-sato/webkit/webkitgtk_2.26.4.bb
@@ -61,7 +61,9 @@ PACKAGECONFIG[openjpeg] = "-DUSE_OPENJPEG=ON,-DUSE_OPENJPEG=OFF,openjpeg"
 # webkitgtk is full of /usr/bin/env python, particular for generating docs
 do_configure[postfuncs] += "setup_python_link"
 setup_python_link() {
-	ln -s `which python3` ${STAGING_BINDIR_NATIVE}/python
+	if [ ! -e ${STAGING_BINDIR_NATIVE}/python ]; then
+		ln -s `which python3` ${STAGING_BINDIR_NATIVE}/python
+	fi
 }
 
 EXTRA_OECMAKE = " \
